+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-24T09:26:51.016Z",
+    "2026-03-24T14:06:08.710Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {