+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3628,7 +3628,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KMFJPMRNRRWZXGAHWPX4E5NF"),
+  id: z.string().default("01KMG2P2041TFF1WGCM1JNFB16"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),